# ARROW-8692: [C++] Avoid memory copies when downloading from S3