@pitrou 2020-05-04
@lidavidm It would be nice if you could run the benchmarks and post number on your setup (perhaps on S3 too?).