@fsaintjacques 2020-05-04
Locally:
```
# Before
$ time cpp/build/conda-release/release/dataset-parquet-scan-example 's3://123:12345678@nyc-tlc/parquet?scheme=http&endpoint_override=localhost:9000'
Table size: 53747                                                                                                                            
                                                                                                                                             
real    0m6.917s                                                                                                                             
user    0m49.500s                                                                                                                                                                                                                                                                         
sys     0m3.758s    
# After
$ time cpp/build/conda-release/release/dataset-parquet-scan-example 's3://123:12345678@nyc-tlc/parquet?scheme=http&endpoint_override=localhost:9000'
Table size: 53747

real    0m6.456s
user    0m45.141s
sys     0m3.322s
```

A small improvement.