@wesm 2020-05-04
S3 benchmarks run outside of EC2 aren't likely to be useful